+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/auth/controller/OAuthController.java
@@ -4,10 +4,9 @@ import com.hertz.hertz_be.domain.auth.dto.request.OAuthLoginRequestDto;
 import com.hertz.hertz_be.domain.auth.dto.response.OAuthLoginResponseDto;
 import com.hertz.hertz_be.domain.auth.dto.response.OAuthLoginResult;
 import com.hertz.hertz_be.domain.auth.dto.response.OAuthSignupResponseDto;
-import com.hertz.hertz_be.domain.auth.exception.OAuthResponseCode;
+import com.hertz.hertz_be.domain.auth.responsecode.OAuthResponseCode;
 import com.hertz.hertz_be.domain.auth.service.OAuthService;
 import com.hertz.hertz_be.domain.channel.service.ChannelService;
-import com.hertz.hertz_be.global.common.ResponseCode;
 import com.hertz.hertz_be.global.common.ResponseDto;
 import com.hertz.hertz_be.global.util.AuthUtil;
 import io.swagger.v3.oas.annotations.Operation;
