+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/auth/service/AuthService.java
@@ -1,9 +1,9 @@
 package com.hertz.hertz_be.domain.auth.service;
 
 import com.hertz.hertz_be.domain.auth.dto.response.ReissueAccessTokenResponseDto;
-import com.hertz.hertz_be.domain.auth.exception.AuthResponseCode;
+import com.hertz.hertz_be.domain.auth.responsecode.AuthResponseCode;
 import com.hertz.hertz_be.domain.auth.repository.RefreshTokenRepository;
-import com.hertz.hertz_be.domain.user.exception.UserResponseCode;
+import com.hertz.hertz_be.domain.user.responsecode.UserResponseCode;
 import com.hertz.hertz_be.domain.user.repository.UserRepository;
 import com.hertz.hertz_be.global.auth.token.JwtTokenProvider;
 import com.hertz.hertz_be.global.exception.BusinessException;
