+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/controller/ChannelController.java
@@ -7,9 +7,9 @@ import com.hertz.hertz_be.domain.channel.dto.response.ChannelRoomResponseDto;
 import com.hertz.hertz_be.domain.channel.dto.response.SendSignalResponseDto;
 import com.hertz.hertz_be.domain.channel.dto.response.TuningResponseDto;
 import com.hertz.hertz_be.domain.channel.entity.enums.MatchingStatus;
-import com.hertz.hertz_be.domain.channel.exception.ChannelResponseCode;
+import com.hertz.hertz_be.domain.channel.responsecode.ChannelResponseCode;
 import com.hertz.hertz_be.domain.channel.service.ChannelService;
-import com.hertz.hertz_be.domain.user.exception.UserResponseCode;
+import com.hertz.hertz_be.domain.user.responsecode.UserResponseCode;
 import com.hertz.hertz_be.global.common.NewResponseCode;
 import com.hertz.hertz_be.global.common.ResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
