+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/responsecode/BaseChannelException.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.domain.channel.exception;
+package com.hertz.hertz_be.domain.channel.responsecode;
 
 public abstract class BaseChannelException extends RuntimeException{
     public abstract String getCode();
