+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/service/ChannelService.java
@@ -3,7 +3,7 @@ package com.hertz.hertz_be.domain.channel.service;
 import com.corundumstudio.socketio.SocketIOServer;
 import com.hertz.hertz_be.domain.channel.dto.request.SignalMatchingRequestDto;
 import com.hertz.hertz_be.domain.channel.dto.response.*;
-import com.hertz.hertz_be.domain.user.exception.UserResponseCode;
+import com.hertz.hertz_be.domain.user.responsecode.UserResponseCode;
 import com.hertz.hertz_be.global.common.NewResponseCode;
 import com.hertz.hertz_be.global.exception.*;
 import com.hertz.hertz_be.global.socketio.dto.SocketIoMessageResponse;
@@ -11,7 +11,7 @@ import com.hertz.hertz_be.domain.channel.entity.*;
 import com.hertz.hertz_be.domain.channel.entity.enums.Category;
 import com.hertz.hertz_be.domain.channel.entity.enums.MatchingStatus;
 import com.hertz.hertz_be.domain.channel.dto.request.SendSignalRequestDto;
-import com.hertz.hertz_be.domain.channel.exception.*;
+import com.hertz.hertz_be.domain.channel.responsecode.*;
 import com.hertz.hertz_be.domain.channel.repository.*;
 import com.hertz.hertz_be.domain.channel.repository.projection.ChannelRoomProjection;
 import com.hertz.hertz_be.domain.channel.repository.projection.RoomWithLastSenderProjection;
