+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/interests/service/InterestsService.java
@@ -9,12 +9,12 @@ import com.hertz.hertz_be.domain.interests.entity.InterestsCategory;
 import com.hertz.hertz_be.domain.interests.entity.InterestsCategoryItem;
 import com.hertz.hertz_be.domain.interests.entity.UserInterests;
 import com.hertz.hertz_be.domain.interests.entity.enums.InterestsCategoryType;
-import com.hertz.hertz_be.domain.interests.exception.*;
+import com.hertz.hertz_be.domain.interests.responsecode.*;
 import com.hertz.hertz_be.domain.interests.repository.InterestsCategoryItemRepository;
 import com.hertz.hertz_be.domain.interests.repository.InterestsCategoryRepository;
 import com.hertz.hertz_be.domain.interests.repository.UserInterestsRepository;
 import com.hertz.hertz_be.domain.user.entity.User;
-import com.hertz.hertz_be.domain.user.exception.UserResponseCode;
+import com.hertz.hertz_be.domain.user.responsecode.UserResponseCode;
 import com.hertz.hertz_be.domain.user.repository.UserRepository;
 import com.hertz.hertz_be.global.common.NewResponseCode;
 import com.hertz.hertz_be.global.exception.BusinessException;
