+++ hertz-be/src/main/java/com/hertz/hertz_be/global/auth/filter/SseAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.hertz.hertz_be.global.auth.filter;
 
-import com.hertz.hertz_be.domain.auth.exception.AuthResponseCode;
+import com.hertz.hertz_be.domain.auth.responsecode.AuthResponseCode;
 import com.hertz.hertz_be.global.auth.token.JwtTokenProvider;
 import com.hertz.hertz_be.global.exception.BusinessException;
 import jakarta.servlet.FilterChain;
